requirements.txt
blinker==1.9.0
boto3==1.38.10
botocore==1.38.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Login==0.6.3
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
qrcode==8.2
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
urllib3==1.26.20
Werkzeug==3.1.3
zipp==3.23.0
